Use climain.js file for bin script
climain.js can be run by nodejs un-transpiled.
climain.js gets copied to dist when publishing as tsdx does not allow multiple entry points.

Moved zenodo_upload() to own file to prevent circular dependencies
