+++ src/climain.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+
+const zu = require('@iomeg/zenodo-upload');
+zu.climain();
